legal: Add a license
I didn't think any of this was sufficiently copyrightable, but the
lawyers asked for a license, so here it is. I was trying to avoid
creating a divide between the different license sects, but I settled on
this because it will surely never be combined with any other software,
and on the odd chance we write any significant amount of code, it might
as well stay free.